Rename HeroNavMenu module and update imports
Moved HeroNavMenu from widgets to features, reflecting a change in its categorization. Updated all relevant imports to the new path and commented out an unused import in picture-galleries layout.
