+++ frontend-next-migration/src/features/HeroNavMenu/ui/HeroNavMenu.tsx
